+++ lomond/errors.py
@@ -20,6 +20,13 @@ class HandshakeError(WebSocketError):
     """
 
 
+class CompressionParameterError(HandshakeError):
+    """
+    Raised when the headers contain invalid compression parameters.
+
+    """
+
+
 class ProtocolError(WebSocketError):
     """Raised in response to a protocol violation."""
     # Results in a a graceful disconnect.