+++ lomond/parser.py
@@ -60,9 +60,7 @@ class _ReadUntil(_Awaitable):
         """Check the length is within max bytes."""
         if self.max_bytes is not None and pos > self.max_bytes:
             raise ParseError(
-                '{!r} expected in {} byte(s)',
-                self.sep,
-                self.max_bytes
+                'expected {!r}'.format(self.sep)
             )
 
 
@@ -90,7 +88,7 @@ class Parser(object):
         self._awaiting = None
         self._buffer = []  # Buffer for reads
         self._until = b''  # Buffer for read untils
-        self._closed = False
+        self._eof = False
         self.reset()
 
     read = _ReadBytes
@@ -99,6 +97,10 @@ class Parser(object):
 
     def __del__(self):
         self.close()
+
+    @property
+    def is_eof(self):
+        return self._eof
 
     def reset(self):
         """Reset the parser, so it may be used on a fresh stream."""
@@ -116,7 +118,7 @@ class Parser(object):
         Called with data (bytes), will yield 0 or more objects parsed
         from the stream.
 
-        :param bytes data: Data to decode.
+        :param bytes data: Data to parse.
 
         """
 
@@ -124,10 +126,15 @@ class Parser(object):
             try:
                 self._awaiting.check_length(pos)
             except ParseError as error:
-                self._gen.throw(error)
+                self._awaiting = self._gen.throw(error)
 
+        if self._eof:
+            raise ParseError('end of file reached')
         if not data:
-            self._gen.throw(ParseEOF('no more data to parse'))
+            self._eof = True
+            self._gen.throw(
+                ParseError('unexpected eof of file')
+            )
 
         pos = 0
         while pos < len(data):
@@ -144,7 +151,7 @@ class Parser(object):
                     self._awaiting.validate(chunk)
                 except ParseError as error:
                     # Raises an exception in parse()
-                    self._gen.throw(error)
+                    self._awaiting = self._gen.throw(error)
                 # Add to buffer
                 self._buffer.append(chunk)
                 remaining -= chunk_size
@@ -173,15 +180,16 @@ class Parser(object):
                 else:
                     # Found separator
                     # Get data prior to and including separator
-                    _check_length(sep_index + len(sep))
-                    split_index = sep_index + len(sep)
-                    send_bytes = self._until[:split_index]
+                    sep_index += len(sep)
+                    _check_length(sep_index)
+                    send_bytes = self._until[:sep_index]
                     # Reset data, to continue parsing
-                    data = self._until[split_index:]
+                    data = self._until[sep_index:]
                     self._until = b''
                     pos = 0
                     # Send bytes to coroutine, get new 'awaitable'
                     self._awaiting = self._gen.send(send_bytes)
+
             # Yield any non-awaitables...
             while not isinstance(self._awaiting, _Awaitable):
                 yield self._awaiting