+++ lomond/stream.py
@@ -12,7 +12,7 @@ import logging
 from six import text_type
 
 from . import errors
-from .frame_parser import FrameParser
+from .frame_parser import ClientFrameParser
 from .message import Message
 from .parser import ParseError
 from .response import Response
@@ -29,22 +29,36 @@ class WebsocketStream(object):
     """
 
     def __init__(self):
-        self.frame_parser = FrameParser()
+        self.frame_parser = ClientFrameParser()
         self._parsed_response = False
         self._frames = []
+        self._compression = None
+        self._decompress = None
+
+    def set_compression(self, compression):
+        """Set a compression object for decompressing messages."""
+        self.frame_parser.enable_compression()
+        self._compression = compression
+        self._decompress = (
+            self._compression.decompress
+            if self._compression
+            else None
+        )
+
+    def build_message(self, frames):
+        """Return a message, built from a list of frames."""
+        return Message.build(frames, self._decompress)
 
     def feed(self, data):
         """Feed in data from a socket to yield 0 or more frames."""
         # This combines fragmented frames in to a single frame
+
         iter_frames = iter(self.frame_parser.feed(data))
 
         if not self._parsed_response:
-            # Process status line and headers from frame parser
-            header_data = next(iter_frames, None)
-            if header_data is None:
-                return
-            self._parsed_response = True
+            header_data = next(iter_frames)
             yield Response(header_data)
+            self._parsed_response = True
 
         # Process incoming frames
         while True:
@@ -54,11 +68,11 @@ class WebsocketStream(object):
                 raise errors.CriticalProtocolError(
                     text_type(error)
                 )
-            log.debug("SRV -> CLI : %r", frame)
+            log.debug(" SRV -> CLI : %r", frame)
             if frame.is_control:
                 # Control messages are never fragmented
                 # And may be sent in the middle of a multi-part message
-                yield Message.build([frame])
+                yield self.build_message([frame])
             else:
                 # May be fragmented
                 if frame.is_continuation and not self._frames:
@@ -71,5 +85,5 @@ class WebsocketStream(object):
                     )
                 self._frames.append(frame)
                 if frame.fin:
-                    yield Message.build(self._frames)
+                    yield self.build_message(self._frames)
                     del self._frames[:]